Move lib/client/BaseGraphqlPayload.js below lib/client/graphql/
