+++ lib/client/graphql/BaseGraphqlPayload.js
@@ -1,4 +1,4 @@
-import UploadingPropertyPathBuilder from '../domClerks/UploadingPropertyPathBuilder.js'
+import UploadingPropertyPathBuilder from '../../domClerks/UploadingPropertyPathBuilder.js'
 
 /**
  * Base class of GraphQL payload.
