+++ tests/__tests__/app/graphql/client/BaseAppGraphqlPayload.js
@@ -1,5 +1,5 @@
 import BaseAppGraphqlPayload from '~/app/graphql/clients/BaseAppGraphqlPayload.js'
-import BaseGraphqlPayload from '~/lib/client/BaseGraphqlPayload.js'
+import BaseGraphqlPayload from '~/lib/client/graphql/BaseGraphqlPayload.js'
 import StorageClerk from '~/lib/storage/StorageClerk.js'
 
 beforeEach(() => {
