+++ tests/__tests__/lib/client/BaseGraphqlCapsule.js
@@ -6,7 +6,7 @@ import BaseGraphqlCapsule, {
   LAUNCH_ABORTED_REASON,
 } from '~/lib/client/graphql/BaseGraphqlCapsule.js'
 
-import BaseGraphqlPayload from '~/lib/client/BaseGraphqlPayload.js'
+import BaseGraphqlPayload from '~/lib/client/graphql/BaseGraphqlPayload.js'
 
 describe('BaseGraphqlCapsule', () => {
   describe('constructor', () => {
