+++ tests/__tests__/lib/client/BaseGraphqlLauncher.js
@@ -3,7 +3,7 @@ import {
 } from '@openreachtech/jest-constructor-spy'
 
 import BaseGraphqlLauncher from '~/lib/client/graphql/BaseGraphqlLauncher.js'
-import BaseGraphqlPayload from '~/lib/client/BaseGraphqlPayload.js'
+import BaseGraphqlPayload from '~/lib/client/graphql/BaseGraphqlPayload.js'
 import {
   default as BaseGraphqlCapsule,
   LAUNCH_ABORTED_REASON,
